Consolidate schema code and tests into schema.h/schema-internal.h
