+++ src/parquet/column/scanner.h
@@ -27,7 +27,7 @@
 
 #include "parquet/column/reader.h"
 #include "parquet/exception.h"
-#include "parquet/schema/descriptor.h"
+#include "parquet/schema.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
 #include "parquet/util/visibility.h"
