+++ src/parquet/column/statistics.h
@@ -22,7 +22,7 @@
 #include <memory>
 #include <string>
 
-#include "parquet/schema/descriptor.h"
+#include "parquet/schema.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
 #include "parquet/util/visibility.h"
