+++ src/parquet/file/metadata.cc
@@ -21,7 +21,8 @@
 
 #include "parquet/exception.h"
 #include "parquet/file/metadata.h"
-#include "parquet/schema/converter.h"
+#include "parquet/schema-internal.h"
+#include "parquet/schema.h"
 #include "parquet/thrift/util.h"
 #include "parquet/util/memory.h"
