+++ src/parquet/file/metadata.h
@@ -25,7 +25,7 @@
 #include "parquet/column/properties.h"
 #include "parquet/column/statistics.h"
 #include "parquet/compression.h"
-#include "parquet/schema/descriptor.h"
+#include "parquet/schema.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
 #include "parquet/util/visibility.h"
