+++ src/parquet/file/reader-internal.cc
@@ -27,9 +27,7 @@
 #include "parquet/column/page.h"
 #include "parquet/compression.h"
 #include "parquet/exception.h"
-#include "parquet/schema/converter.h"
-#include "parquet/schema/descriptor.h"
-#include "parquet/schema/types.h"
+#include "parquet/schema.h"
 #include "parquet/thrift/util.h"
 #include "parquet/types.h"
 #include "parquet/util/memory.h"
